#1 - 아날로그 시계 리펙토링
- 시침,분침,초침 모듈화
- 불필요하게 분리 되있는 파일 삭제
